refactor(AppComponents): move icons to a separate file
